feat: Add support for Vietnamese language
This commit adds support for the Vietnamese language in the configuration files and scripts. The `vi` language code has been added to the list of supported languages in both the `docusaurus.config.ts` and `config.mts` files. Additionally, the `vi` language has been included in the `LinguineCore` object in the `linguine.mts` file.

This change was made to accommodate the needs of Vietnamese-speaking users who visit the website or use the language translation feature.